+++ packages/plugin-packages/mashroom-vhost-path-mapper/test/middleware/determine_host.test.ts
@@ -11,22 +11,7 @@ describe('determine_host', () => {
             }
         };
 
-        expect(determineHost(req)).toEqual({
-            hostname: 'my-host.at',
-            port: undefined,
-        });
-    });
-
-    it('determines a host with forwarding header',  () => {
-        const req: any = {
-            // hostname differs because of a X-Forwarded-Host header
-            hostname: 'my-host.at',
-            headers: {
-                host: 'my-original-host.at'
-            }
-        };
-
-        expect(determineHost(req)).toEqual({
+        expect(determineHost([], req)).toEqual({
             hostname: 'my-host.at',
             port: undefined,
         });
@@ -40,26 +25,58 @@ describe('determine_host', () => {
             }
         };
 
-        expect(determineHost(req)).toEqual({
+        expect(determineHost([], req)).toEqual({
             hostname: 'my-host.at',
             port: '8080',
         });
     });
 
+    it('determines a host with forwarding header',  () => {
+        const req: any = {
+            hostname: 'my-host.at',
+            headers: {
+                host: 'my-host.at',
+                'x-forwarded-host': 'my-original-host.at'
+            }
+        };
+
+        expect(determineHost(['x-forwarded-host'], req)).toEqual({
+            hostname: 'my-original-host.at',
+            port: undefined,
+        });
+    });
 
     it('determines a host with a port and a forwarding header',  () => {
         const req: any = {
             hostname: 'my-host.at',
             headers: {
-                host: 'my-original-host.at:8080'
+                host: 'my-host.at:8080',
+                'x-forwarded-host': 'my-original-host.at'
             }
         };
 
-        expect(determineHost(req)).toEqual({
-            hostname: 'my-host.at',
+        expect(determineHost(['x-forwarded-host'], req)).toEqual({
+            hostname: 'my-original-host.at',
             port: undefined,
         });
     });
+
+    it('consider the http header priority',  () => {
+        const req: any = {
+            hostname: 'my-host.at',
+            headers: {
+                host: 'my-host.at',
+                'x-my-custom-host-header': ['the-real-original-host.at', 'another-host.com'],
+                'x-forwarded-host': 'my-original-host.at'
+            }
+        };
+
+        expect(determineHost(['x-my-custom-host-header', 'x-forwarded-host'], req)).toEqual({
+            hostname: 'the-real-original-host.at',
+            port: undefined,
+        });
+    });
+
 
 });